requirements.txt
dataclasses-json
streamlit
pandas
python-dotenv
openai
pillow